refactor: PluginService class
