+++ beagle-intellij-plugin-core/src/main/kotlin/br/com/zup/beagle/plugin/service/StopRunning.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package br.com.zup.beagle.plugin.service
+
+import br.com.zup.beagle.plugin.runner.BeagleRunConfiguration
+import com.intellij.execution.impl.ExecutionManagerImpl
+import com.intellij.openapi.project.Project
+
+class StopRunning(private val project: Project) {
+
+    private val executionManager = ExecutionManagerImpl.getInstance(this.project)
+
+    fun stopRunningCustomRunConfigurations() {
+        val descriptors = this.executionManager.getDescriptors {
+            it.configuration is BeagleRunConfiguration
+        }
+        descriptors.forEach {
+            if (!it.processHandler!!.isProcessTerminated) {
+                it.processHandler!!.destroyProcess()
+            }
+        }
+    }
+}
